+++ src/pages/questions/Questions.tsx
@@ -1,4 +1,3 @@
-import QuestionsSideBar from "@/components/questions/QuestionsSideBar.tsx";
 import {GiHand} from "react-icons/gi";
 import {Button} from "@/components/ui/button.tsx";
 import {Input} from "@/components/ui/input.tsx";
@@ -6,36 +5,22 @@ import {BiSearch} from "react-icons/bi";
 import {useFindAllQuestionInfinite} from "@/openapi/question/api/questions/questions.ts";
 import {MdExpandMore} from "react-icons/md";
 import {Badge} from "@/components/ui/badge.tsx";
-import {faker} from '@faker-js/faker';
 import {AiFillLike, AiOutlineComment} from "react-icons/ai";
 import {IoGlasses} from "react-icons/io5";
 import timeSince from "@/components/utils/timeSince.tsx";
-
-
-function createFakeDataArray(dataCount: number){
-    return Array.from({length: dataCount}, () => ({
-        title: faker.lorem.sentence(),
-        content: faker.lorem.paragraphs(),
-        createdAt: faker.date.past().toISOString(),
-        author: faker.person.fullName(), // faker.person.fullName() 대신 faker.name.findName()을 사용
-        tags: Array.from({length: 10}, () => faker.lorem.word()),
-        likes: faker.number.int({min: 0, max: 500}),
-        views: faker.number.int({min: 0, max: 5000}),
-        answers: faker.number.int({min: 0, max: 50})
-    }));
-}
+import {useNavigate} from "react-router-dom";
 
 const Questions = () => {
 
+    const navigate = useNavigate();
+
     const {
-        data: questions,
+        data: questionsDatas,
         fetchNextPage,
         hasNextPage,
         isFetchingNextPage,
     } = useFindAllQuestionInfinite({
-        pageable: {
-
-        }
+        pageable: {}
     }, {
         query: {
             queryKey: ['Questions'],
@@ -48,20 +33,12 @@ const Questions = () => {
             },
         }
     });
+    console.log("questionsDatas", questionsDatas)
 
-    console.log("questions", questions)
-
-    // 가짜 데이터 생성 및 출력
-    const fakeDataArray = createFakeDataArray(10);
-    console.log(fakeDataArray);
+    const questionsData = questionsDatas?.pages.map(page => page.content)
 
     return (
-        <div
-            className="flex flex-1 flex-col mt-[65px] bg-white dark:bg-[#1E1E1E] overflow-y-auto mx-5 sm:mx-[50px] md:mx-[100px] lg:mx-[150px] xl:mx-[250px] 2xl:mx-[400px]">
-
-            {/* 사이드 바 */}
-            <QuestionsSideBar/>
-
+        <>
             <div className="bg-transparent flex flex-1 flex-col p-3">
                 {/* 질문하기 */}
                 <div className="flex justify-end">
@@ -84,57 +61,69 @@ const Questions = () => {
 
                 {/* Q&A 목록 */}
                 <div className="flex flex-1 flex-col justify-start bg-transparent">
-                    {fakeDataArray.map((fakeData, index) => {
-                        return (
-                            <div key={index}
-                                 className="flex flex-col w-full bg-transparent p-2 sm:p-3 border-b border-b-gray-300">
+                    {questionsData?.map((questions) => (
+                        questions?.map((question, index) => {
+                            return (
+                                <div
+                                    onClick={() => {
+                                        navigate(`/questions/${question.questionId}`)
+                                    }}
+                                    key={index}
+                                    className="flex flex-col w-full bg-transparent p-2 sm:p-5 border-b border-b-gray-300 hover:bg-gray-100 dark:hover:bg-neutral-900 cursor-pointer">
 
-                                <div className="flex flex-col">
-                                    <div className="text-xl font-semibold">{fakeData.title}</div>
-                                    <div>{fakeData.content?.substring(0, 100)}...</div>
-                                </div>
+                                    <div className="flex flex-col">
+                                        <div className="text-lg font-semibold">{question?.title}</div>
+                                        <div className="text-sm">{question?.content?.substring(0, 100)}...</div>
+                                    </div>
 
-                                <div className="mt-2">
-                                    <div className="hidden sm:flex">
-                                        {fakeData.tags.map((tag: string) => {
-                                            return (
-                                                <div className="">
-                                                    {tag.length <= 9 &&
-                                                        <Badge
-                                                            className="text-white bg-indigo-300 hover:bg-indigo-400 dark:bg-indigo-500 dark:hover:bg-indigo-600 mr-1">{tag}</Badge>
-                                                    }
+                                    <div className="mt-3">
+                                        <div className="hidden sm:flex">
+                                            {question?.tags?.tags?.map((tag: string) => {
+                                                return (
+                                                    <>
+                                                        {tag.length <= 9 &&
+                                                            <Badge
+                                                                className="mt-3 text-white bg-indigo-300 hover:bg-indigo-400 dark:bg-indigo-500 dark:hover:bg-indigo-600 mr-1">{tag}</Badge>
+                                                        }
+                                                    </>
+                                                );
+                                            })}
+                                        </div>
+
+                                        <div className="flex items-center justify-between mt-1">
+                                            <div className="flex text-xs space-x-2">
+                                                <div>{question?.author?.username || question?.author?.nickname}</div>
+                                                <div
+                                                    className="text-gray-500 dark:text-gray-400">
+                                                    {question.createdAt && timeSince(new Date(question.createdAt))}
                                                 </div>
-                                            );
-                                        })}
-                                    </div>
+                                            </div>
 
-                                    <div className="flex items-center justify-between mt-1 px-1">
-                                        <div className="flex text-sm justify-end space-x-1">
-                                            <div>{fakeData.author}</div>
-                                            <div
-                                                className="text-gray-500 dark:text-gray-400">{timeSince(new Date(fakeData.createdAt))}</div>
-                                        </div>
-                                        <div className="flex text-sm space-x-1.5">
-                                            <div className="flex items-center space-x-0.5">
-                                                <AiFillLike className="w-4 h-4"/>
-                                                <span>{fakeData.likes}</span>
-                                            </div>
-                                            <div className="text-gray-400">|</div>
-                                            <div className="flex items-center space-x-0.5">
-                                                <IoGlasses className="w-6 h-6"/>
-                                                <span>{fakeData.views}</span>
-                                            </div>
-                                            <div className="text-gray-400">|</div>
-                                            <div className="flex items-center space-x-0.5">
-                                                <AiOutlineComment className="w-5 h-5"/>
-                                                <span>{fakeData.answers}</span>
+                                            <div className="flex text-xs space-x-1.5">
+                                                <div
+                                                    className="flex items-center space-x-0.5 text-gray-500 dark:text-gray-400">
+                                                    <AiFillLike className="w-3.5 h-3.5"/>
+                                                    <span>미</span>
+                                                </div>
+                                                <div className="text-gray-400">|</div>
+                                                <div
+                                                    className="flex items-center space-x-0.5 text-gray-500 dark:text-gray-400">
+                                                    <IoGlasses className="w-5 h-5"/>
+                                                    <span>구</span>
+                                                </div>
+                                                <div className="text-gray-400">|</div>
+                                                <div
+                                                    className="flex items-center space-x-0.5 text-gray-500 dark:text-gray-400">
+                                                    <AiOutlineComment className="w-4 h-4"/>
+                                                    <span>현</span>
+                                                </div>
                                             </div>
                                         </div>
                                     </div>
                                 </div>
-                            </div>
-                        )
-                    })}
+                            )
+                        })
+                    ))}
                 </div>
             </div>
 
@@ -147,7 +136,7 @@ const Questions = () => {
                     </Button>
                 </div>
             )}
-        </div>
+        </>
     )
 }
